*: add metrics for unified read pool (#6534)
