+++ Cargo.toml
@@ -131,17 +131,16 @@ vlog = "0.1.4"
 git = "https://github.com/pingcap/murmur3.git"
 
 [dependencies.prometheus]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 features = ["nightly", "push", "process"]
 
 [dependencies.prometheus-static-metric]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 
 [dependencies.yatp]
 git = "https://github.com/tikv/yatp.git"
-rev = "4e3ac06b948f1c2dffe92986a760ef561d0bdbb9"
 
 [dev-dependencies]
 panic_hook = { path = "components/panic_hook" }
