+++ cmd/src/server.rs
@@ -29,23 +29,21 @@ use std::{
     thread::JoinHandle,
     time::Duration,
 };
-use tikv::config::ConfigHandler;
-use tikv::raftstore::coprocessor::config::SplitCheckConfigManager;
-use tikv::raftstore::router::ServerRaftStoreRouter;
-use tikv::raftstore::store::config::RaftstoreConfigManager;
 use tikv::{
-    config::{ConfigController, DBConfigManger, DBType, TiKvConfig},
+    config::{ConfigController, ConfigHandler, DBConfigManger, DBType, TiKvConfig},
     coprocessor,
     import::{ImportSSTService, SSTImporter},
     raftstore::{
-        coprocessor::{CoprocessorHost, RegionInfoAccessor},
+        coprocessor::{config::SplitCheckConfigManager, CoprocessorHost, RegionInfoAccessor},
+        router::ServerRaftStoreRouter,
         store::{
+            config::RaftstoreConfigManager,
             fsm,
             fsm::store::{RaftBatchSystem, RaftRouter, StoreMeta, PENDING_VOTES_CAP},
             new_compaction_listener, LocalReader, PdTask, SnapManagerBuilder, SplitCheckRunner,
         },
     },
-    read_pool::{ReadPool, ReadPoolRunner},
+    read_pool::{build_yatp_read_pool, ReadPool},
     server::{
         config::Config as ServerConfig,
         create_raft_storage,
@@ -65,10 +63,6 @@ use tikv_util::{
     time::Monitor,
     worker::{FutureScheduler, FutureWorker, Worker},
 };
-use yatp::{
-    pool::CloneRunnerBuilder,
-    queue::{multilevel, QueueType},
-};
 
 /// Run a TiKV server. Returns when the server is shutdown by the user, in which
 /// case the server will be properly stopped.
@@ -81,6 +75,7 @@ pub fn run_tikv(config: TiKvConfig) {
     let _m = Monitor::default();
 
     tikv.init_fs();
+    tikv.init_yatp();
     tikv.init_engines();
     let gc_worker = tikv.init_gc_worker();
     let server_config = tikv.init_servers(&gc_worker);
@@ -281,6 +276,17 @@ impl TiKVServer {
         .unwrap();
     }
 
+    fn init_yatp(&self) {
+        let yatp_enabled = self.config.readpool.unify_read_pool;
+        if !yatp_enabled {
+            return;
+        }
+
+        yatp::metrics::set_namespace(Some("tikv"));
+        prometheus::register(Box::new(yatp::metrics::MULTILEVEL_LEVEL0_CHANCE.clone())).unwrap();
+        prometheus::register(Box::new(yatp::metrics::MULTILEVEL_LEVEL_ELAPSED.clone())).unwrap();
+    }
+
     fn init_engines(&mut self) {
         let block_cache = self.config.storage.block_cache.build_shared_cache();
 
@@ -403,18 +409,10 @@ impl TiKVServer {
         let pd_sender = pd_worker.scheduler();
 
         let unified_read_pool = if self.config.readpool.unify_read_pool {
-            let unified_read_pool_cfg = &self.config.readpool.unified;
-            let mut builder = yatp::Builder::new("unified-read-pool");
-            builder
-                .min_thread_count(unified_read_pool_cfg.min_thread_count)
-                .max_thread_count(unified_read_pool_cfg.max_thread_count);
-            let multilevel_builder = multilevel::Builder::new(Default::default());
-            let read_pool_runner = ReadPoolRunner::new(engines.engine.clone(), Default::default());
-            let runner_builder =
-                multilevel_builder.runner_builder(CloneRunnerBuilder(read_pool_runner));
-            Some(builder.build_with_queue_and_runner(
-                QueueType::Multilevel(multilevel_builder),
-                runner_builder,
+            Some(build_yatp_read_pool(
+                &self.config.readpool.unified,
+                pd_sender.clone(),
+                engines.engine.clone(),
             ))
         } else {
             None
@@ -423,12 +421,12 @@ impl TiKVServer {
         let storage_read_pool = if self.config.readpool.unify_read_pool {
             ReadPool::from(unified_read_pool.as_ref().unwrap().remote().clone())
         } else {
-            let cop_read_pools = storage::build_read_pool(
+            let storage_read_pools = storage::build_read_pool(
                 &self.config.readpool.storage,
                 pd_sender.clone(),
                 engines.engine.clone(),
             );
-            ReadPool::from(cop_read_pools)
+            ReadPool::from(storage_read_pools)
         };
 
         let storage = create_raft_storage(
