+++ components/backup/Cargo.toml
@@ -39,7 +39,7 @@ tokio-threadpool = "0.1"
 url = "2.0"
 
 [dependencies.prometheus]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 default-features = false
 features = ["nightly", "push", "process"]
