+++ components/engine/Cargo.toml
@@ -30,7 +30,7 @@ tikv_util = { path = "../tikv_util" }
 time = "0.1"
 
 [dependencies.prometheus]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 features = ["nightly", "push", "process"]
