+++ components/pd_client/Cargo.toml
@@ -27,7 +27,7 @@ tokio-timer = "0.2"
 txn_types = { path = "../txn_types" }
 
 [dependencies.prometheus]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 features = ["nightly", "push", "process"]
