+++ components/tidb_query/Cargo.toml
@@ -52,17 +52,16 @@ uuid = { version = "0.8.1", features = ["v4"] }
 static_assertions = { version = "1.0", features = ["nightly"] }
 
 [dependencies.prometheus]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 features = ["nightly", "push", "process"]
 
 [dependencies.prometheus-static-metric]
-git = "https://github.com/pingcap/rust-prometheus.git"
+git = "https://github.com/tikv/rust-prometheus.git"
 rev = "d919ccd35976b9b84b8d03c07138c1cc05a36087"
 
 [dependencies.yatp]
 git = "https://github.com/tikv/yatp.git"
-rev = "4e3ac06b948f1c2dffe92986a760ef561d0bdbb9"
 
 [dev-dependencies]
 profiler = { path = "../profiler" }
